Assignment_7/donut-model-demo/simple_requirements.txt
torch
transformers
pillow
protobuf
sentencepiece
pdf2image
